requirements.txt
discord.py-self==2.0.0
requests==2.31.0
colorama==0.4.6